+++ test/text/test_render_styled.py
@@ -11,38 +11,46 @@ HERE = Path(__file__).resolve().parent
 
 
 CELLDATA = (
-    # txt, align, new_x, new_y
+    # txt,        align, new_x,    new_y
     ("Left Top L", "L", "LEFT", "TOP"),
     ("Left Top R", "R", "LEFT", "TOP"),
     ("Left Top C", "C", "LEFT", "TOP"),
+    ("Left Top X", "X", "LEFT", "TOP"),
     ("Left Top J", "J", "LEFT", "TOP"),
     ("Right Last L", "L", "RIGHT", "LAST"),
     ("Right Last R", "R", "RIGHT", "LAST"),
     ("Right Last C", "C", "RIGHT", "LAST"),
+    ("Right Last X", "X", "RIGHT", "LAST"),
     ("Right Last J", "J", "RIGHT", "LAST"),
     ("Start Next L", "L", "START", "NEXT"),
     ("Start Next R", "R", "START", "NEXT"),
     ("Start Next C", "C", "START", "NEXT"),
+    ("Start Next X", "X", "START", "NEXT"),
     ("Start Next J", "J", "START", "NEXT"),
     ("End TMargin L", "L", "END", "TMARGIN"),
     ("End TMargin R", "R", "END", "TMARGIN"),
     ("End TMargin C", "C", "END", "TMARGIN"),
+    ("End TMargin X", "X", "END", "TMARGIN"),
     ("End TMargin J", "J", "END", "TMARGIN"),
     ("WCont Top L", "L", "WCONT", "TOP"),
     ("WCont Top R", "R", "WCONT", "TOP"),
     ("WCont Top C", "C", "WCONT", "TOP"),
+    ("WCont Top X", "X", "WCONT", "TOP"),
     ("WCont Top J", "J", "WCONT", "TOP"),
     ("Center TOP L", "L", "CENTER", "TOP"),
     ("Center TOP R", "R", "CENTER", "TOP"),
     ("Center TOP C", "C", "CENTER", "TOP"),
+    ("Center TOP X", "X", "CENTER", "TOP"),
     ("Center TOP J", "J", "CENTER", "TOP"),
     ("LMargin BMargin L", "L", "LMARGIN", "BMARGIN"),
     ("LMargin BMargin R", "R", "LMARGIN", "BMARGIN"),
     ("LMargin BMargin C", "C", "LMARGIN", "BMARGIN"),
+    ("LMargin BMargin X", "X", "LMARGIN", "BMARGIN"),
     ("LMargin BMargin J", "J", "LMARGIN", "BMARGIN"),
     ("RMargin Top L", "L", "RMARGIN", "TOP"),
     ("RMargin Top R", "R", "RMARGIN", "TOP"),
     ("RMargin Top C", "C", "RMARGIN", "TOP"),
+    ("RMargin Top X", "X", "RMARGIN", "TOP"),
     ("RMargin Top J", "J", "RMARGIN", "TOP"),
 )
 
@@ -55,14 +63,13 @@ def test_render_styled_newpos(tmp_path):
     """
     doc = fpdf.FPDF()
     doc.set_font("helvetica", style="U", size=24)
-    doc.set_margin(10)
     twidth = 100
 
     for i, item in enumerate(CELLDATA):
-        i = i % 4
+        i = i % 5
         if i == 0:
             doc.add_page()
-        doc.x = 20
+        doc.x = 70
         doc.y = 20 + (i * 20)
         s = item[0]
         align = Align.coerce(item[1])
@@ -112,14 +119,13 @@ def test_cell_newpos(tmp_path):
     """
     doc = fpdf.FPDF()
     doc.set_font("helvetica", style="U", size=24)
-    doc.set_margin(10)
     twidth = 100
 
     for i, item in enumerate(CELLDATA):
-        i = i % 4
+        i = i % 5
         if i == 0:
             doc.add_page()
-        doc.x = 20
+        doc.x = 70
         doc.y = 20 + (i * 20)
         s = item[0]
         align = item[1]
@@ -156,14 +162,13 @@ def test_multi_cell_newpos(tmp_path):
     """
     doc = fpdf.FPDF()
     doc.set_font("helvetica", style="U", size=24)
-    doc.set_margin(10)
     twidth = 100
 
     for i, item in enumerate(CELLDATA):
-        i = i % 4
+        i = i % 5
         if i == 0:
             doc.add_page()
-        doc.x = 20
+        doc.x = 70
         doc.y = 20 + (i * 20)
         s = item[0]
         align = item[1]
@@ -218,7 +223,6 @@ def test_cell_lnpos(tmp_path):
     """
     doc = fpdf.FPDF()
     doc.set_font("helvetica", style="U", size=24)
-    doc.set_margin(10)
     twidth = 100
 
     for i, item in enumerate(LN_CELLDATA):
@@ -263,7 +267,6 @@ def test_multi_cell_ln_newpos(tmp_path):
     """
     doc = fpdf.FPDF()
     doc.set_font("helvetica", style="U", size=24)
-    doc.set_margin(10)
     twidth = 100
 
     for i, item in enumerate(LN_CELLDATA):